% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarcassRecovery.R
\name{CarcassDetection}
\alias{CarcassDetection}
\title{CarcassRecovery}
\usage{
CarcassRecovery(Abund, CV.Abund, carcassDetected, ASR, removals, propResidency)
}
\arguments{
\item{df}{an object of class 'data.frame'}
}
\description{
Estimate the fraction of cetacean carcasses documented.
}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
